Add ir_moveinto_wedge method to BrillouinZone
For symmetrising, e.g., inelastic neutron scattering data which depends 
on underlying crystal symmetries *and* instrumental effects, it can be 
useful to determine a symmetry-equivalent Q with the same |Q| that lies 
within the irreducible reciprocal wedge. The new method finds such an 
equivalent Qᵢ and the transformation matrix R for all input Q such that 
Q = R*Qᵢ.

Fixed an error with irreducible wedge determination from a 
volume-satisfying irreducible polyhedron. In the case of an 
inversionless pointgroup the found irP can be comprised of two convex 
polyhedra which are each others inverse. The old method would then 
include for each wedge normal from the first an opposite normal from the 
second and no point could be classified as 'in' the wedge. Now the 
method looks for opposite wedge normals and only includes the first of 
any pair -- this will fail if the two convex polyhedra are somehow 
interleaved.

The InterpolationData method replace_data has two new templated 
overloaded versions to attempt automatic conversion of integer types in 
the std::array<INTEGER,3> element specification.
Typo in interpolate_at fixed which caused an out-of-bounds memory access 
when interpolating stored matrices.