Re-arranged interpolation_data:replace_data methods
A problem in commit 675401aa9494ed23dec896a9ecbe68e5ff4bf385 led to the
MSVS compiler complaining about no known conversion between
std::array<size_t,3> and, effectively, std::array<unsigned long,3> in
calls to interpolation_data:replace_data.
The 'fix' was to add templated overloads to handle the array element
type conversion when necessary. One (or possibly two) errors in the fix
led to the observed hung behaviour when replace_data was called.

The first error was failing to pass the converted std::array to the
original version of replace_data, and instead the templated overload was
called indefinitely.

The second error, which may not have caused problems at this point, was
in failing to provide an appropriate specialized form of the templated
functions.

This commit fixes both of the errors from the previous fix.