Add random points in Polyhedron, update tests
A very irritating bug popped up whereby a previously-passing BZTrellis
test stopped working following an interactive rebase which squashed a
number of commits. The test continued to work when compiled with GCC but
failed under MSVC. It appears that the actual culprit was the *version*
of MSVC libraries being used to compile the test, as Visual Studio was
updated between the squash and final pre-merge testing. The behaviour
which caused the test to fail was rounding of 0.5, which GCC and
old-MSVC round to 0. while new-MSVC rounds to 1.
This difference in rounding caused interpolation points at the
irreducible Brillouin zone boundary (with primitive lattice unit
components of 0.5) to be assigned first to equivalent first-Bz points
near -0.5 (tau components of 1 instead of 0) and then a symmetry
operation was found to map them back to where they started.
If the interpolated data followed the symmetry of the Spacegroup then
this would not be a problem, so in the case of real data/simulations
this behaviour is fine.
The problem is that the test interpolates between the coordinates of the
BZTrellis vertices (expressed in inverse Angstrom) which certainly *do
not* follow the symmetry of the Spacegroup. So the symmetry element
would rotate the interpolated vector to have the 'wrong' value.
The solution to this testing problem was to change the test. Now it
interpolates at random points distributed within the irreducible
Brillouin zone instead of at points randomly between the vertices of the
Trellis -- this does not remove the possibility of the rounding issue
causing a test fail, but it certainly reduces it significantly. The test
can be further modified to remove any random points which have primitive
lattice unit components which round to 1.

The new test requires a set of random points within the irreducible
Brillouin zone polyhedron. To simplify this task and support, e.g.,
integration of functions over the irBz the Polyhedron class can now
produce a random distribution of points within each (convex) polyhedra.
This method takes a number of points and an optional seed for the random
number generator.