11 hall symbol (#15)
A number of changes, summarized below, have been made to enable
specification of spacegroup directly by the Hall symbol (expressed as a
string of ASCII characters).
Along with these changes, one can now specify the 'choice' of a
non-standard spacegroup setting along with its International Tables
abbreviated name. E.g., to differentiate between Hall symbols 'P 2 2 -1n'
and '-P 2ab 2bc' which are both identified in the International Tables
as 'Pnnn' one could give the choice in addition as '1' or '2',
respectively.

Hall symbol details:
--------------------

A new class, HallSymbol, is able to decode Hall's notation into a set of
generators returned as a Symmetry object. It has been added to the
pybind11 interface whereby it can be created from an ASCII string, and
its generators can be obtained.
Holding the Hall-notation Seitz matrix symbols has required the creation
of another new class, SeitzSymbols, which does not have a python
interface, and might not ever need to be exposed to users.

The Symmetry class contains a new method to generate a full Spacegroup
symmetry from a contained set of generators.
It has been extended to make use of a new Motions class, which is a
specialized implementation of Seitz matrices where the last row of the
4x4 matrix is always [0 0 0 1] implicitly.
If a Symmetry object contains all of the Motions of a full Spacegroup
then it can be used to determine a set of generators for that Spacegroup
as well.
Going from full Spacegroup (Symmetry object) to HallSymbol *may* be
possible programmatically but might prove difficult since the generators
of a Spacegroup are not necessarily unique (and Hall symbols must
contain the minimum set of generators ordered in such a way to take
advantage of Hall's implicit notation as much as possible).

The ternary/duodecimal encoded spacegroup operations from Spglib are
difficult to comprehend and are unmaintainable in the source. Their
origin is probably a script, which is not part of the Spglib repository.
Since the HallSymbol class can now decode the Hall symbols in the Spglib
ALL_SPACEGROUPS array the pre-generated symmetry operations can be
replaced with runtime calculations of full spacegroups from their
generators.

Equivalency between Hall symbol sourced Symmetry objects and those
obtained from the Spglib 'database' is now tested in the C++ tests under
the catch2 tag [HallSymbol]. The Spglib database has been moved to the
test and will no longer be part of the Python or C++ libraries.
This new test passes for all 530 Hall symbols known by Spglib.
For future reference, these are also the 530 Hall symbols in International
Tables of Crystallography Table A1.4.2.7 in Volume B Chapter 1.4

The runtime calculation of Symmetry space groups is probably slower than
decoding from the Spglib list of symmetry operations, but simple testing shows
that it is almost-certainly fast enough. Finding the symmetry operations
at runtime has the added benefit of supporting non-standard Hall symbols.

In order to support non-standard Hall symbols or specification of
the generators by Seitz matrices in 'xyz' notation the Lattice now holds
its spacegroup Symmetry elements (and pointgroup PointSymmetry elements too).
This requires also holding the information derived from the Hall number,
that is the Spacegroup and Pointgroup objects -- which might be left at
default values if the symmetry information is not from a Hall number.

General bugfixes:
-----------------

MSVC is not entirely compliant in its implementation of templates. My
workaround for InterpolationData.replace_data was not acceptable for GCC
since it included a non-namespaced template complete specialization.
Now the main replace_data is templated and a (possibly non-op) type
conversion is always performed. This required making a templated version
of branch_span too.

Fixed an issue using frobenius_distance caused by switching from data
array element specifications of matrix-size to matrix-element-count.
This caused an out-of-bounds memory access by treating a 3x3 matrix as a
9x9 matrix.

* Add random points in Polyhedron, update tests
-------------------------------------------------------

A very irritating bug popped up whereby a previously-passing BZTrellis
test stopped working following an interactive rebase which squashed a
number of commits. The test continued to work when compiled with GCC but
failed under MSVC. It appears that the actual culprit was the *version*
of MSVC libraries being used to compile the test, as Visual Studio was
updated between the squash and final pre-merge testing. The behaviour
which caused the test to fail was rounding of 0.5, which GCC and
old-MSVC round to 0. while new-MSVC rounds to 1.
This difference in rounding caused interpolation points at the
irreducible Brillouin zone boundary (with primitive lattice unit
components of 0.5) to be assigned first to equivalent first-Bz points
near -0.5 (tau components of 1 instead of 0) and then a symmetry
operation was found to map them back to where they started.
If the interpolated data followed the symmetry of the Spacegroup then
this would not be a problem, so in the case of real data/simulations
this behaviour is fine.
The problem is that the test interpolates between the coordinates of the
BZTrellis vertices (expressed in inverse Angstrom) which certainly *do
not* follow the symmetry of the Spacegroup. So the symmetry element
would rotate the interpolated vector to have the 'wrong' value.
The solution to this testing problem was to change the test. Now it
interpolates at random points distributed within the irreducible
Brillouin zone instead of at points randomly between the vertices of the
Trellis -- this does not remove the possibility of the rounding issue
causing a test fail, but it certainly reduces it significantly. The test
can be further modified to remove any random points which have primitive
lattice unit components which round to 1.

The new test requires a set of random points within the irreducible
Brillouin zone polyhedron. To simplify this task and support, e.g.,
integration of functions over the irBz the Polyhedron class can now
produce a random distribution of points within each (convex) polyhedra.
This method takes a number of points and an optional seed for the random
number generator.