Changes for MacOS clang
In an attempt to make brille and MATLAB work together on MacOS machines a MacOS-only instruction was inserted to statically link libaries for the python module.
This instruction is not compatible with clang, and it is the cause of failing Github Action building.
For now brille will continue to be dynamically linked, which may cause problems when calling the brille python module from MATLAB.
The two possible solutions to this are:
	1) dynamically link to the libraries distributed along with MATLAB
	2) compile a statically linked C++ brille library for the python module to use
The module created via option 1 may not be distributable, and would require building brille for specific versions of MATLAB. This may be feasible if CMake can find the MATLAB-included libraries and we ask users to build brille from source.
The second option would better enforce the separation between the brille library and the brille module, and would have the nice feature that a distributable static (and shared, if we want) brille library would be produced. However, there may still be compatibility issues between brille and MATLAB.

The change to vector_cross_impl in utilities.hpp addresses three unused parameter warnings which only clang seems to find.