Fixes for warnings/errors on Action hosts
CTest sometimes requires that the configuration used in building is 
passed to it as, e.g., '-C Release'. This was not included previously in 
the CI script.

Some compilers incorrectly warn about a possibly uninitialised array in 
one hall_symbol.cpp method. Initialising the direction vector to 0 
shouldn't hurt performance and will remove the warning. In fact the 
vector was never used if not set to something sensible but some 
compilers are not clever enough to figure that out.

The minimum version of pybind11 was too low for macOS builds. v2.2.0 
contains something which causes the build to fail. The most recent 
version v2.4.3 works on our mac Mini. There might be versions between 
the two which work as well (v2.2.1, v2.2.2, v2.2.3, v2.2.4, v2.3.0, 
v2.4.0, v2.4.1, v2.4.2) but they have not been tested.
For now set the minimum version to v2.4.3.