Fixes for Actions hosts (#18)
* Changes for MacOS clang

In an attempt to make brille and MATLAB work together on MacOS machines a MacOS-only instruction was inserted to statically link libaries for the python module.
This instruction is not compatible with clang, and it is the cause of failing Github Action building.
For now brille will continue to be dynamically linked, which may cause problems when calling the brille python module from MATLAB.
The two possible solutions to this are:
	1) dynamically link to the libraries distributed along with MATLAB
	2) compile a statically linked C++ brille library for the python module to use
The module created via option 1 may not be distributable, and would require building brille for specific versions of MATLAB. This may be feasible if CMake can find the MATLAB-included libraries and we ask users to build brille from source.
The second option would better enforce the separation between the brille library and the brille module, and would have the nice feature that a distributable static (and shared, if we want) brille library would be produced. However, there may still be compatibility issues between brille and MATLAB.

The change to vector_cross_impl in utilities.hpp addresses three unused parameter warnings which only clang seems to find.

* Move to full Catch2 (fetching if necessary)

By using the full Catch2 we can register individual tests with CTest, 
this will help narrow-down test failures in the remote continuous 
integration.

Two catch2 tests have been disabled since they caused segmentation 
faults on a Hyper-V Windows 10 development VM. These tests were related 
to currently unused functionality which either needs to be removed 
entirely or modified to work with the newest 'grid' classes anyway.

* Fix CMake logic syntax error, remove lib/catch.hpp

* Re-enable Windows MSVC and macOS CI builds

* Remove updated cmake, ninja, and unused gitpython

* Fixes for warnings/errors on Action hosts

CTest sometimes requires that the configuration used in building is 
passed to it as, e.g., '-C Release'. This was not included previously in 
the CI script.

Some compilers incorrectly warn about a possibly uninitialised array in 
one hall_symbol.cpp method. Initialising the direction vector to 0 
shouldn't hurt performance and will remove the warning. In fact the 
vector was never used if not set to something sensible but some 
compilers are not clever enough to figure that out.

The minimum version of pybind11 was too low for macOS builds. v2.2.0 
contains something which causes the build to fail. The most recent 
version v2.4.3 works on our mac Mini. There might be versions between 
the two which work as well (v2.2.1, v2.2.2, v2.2.3, v2.2.4, v2.3.0, 
v2.4.0, v2.4.1, v2.4.2) but they have not been tested.
For now set the minimum version to v2.4.3.