Add search for convex irreducible polyhedra when lacking inversion
Before this change all spacegroups which lack inversion symmetry
produced irreducible polyhedra which were not convex. Instead they were
comprised of a convex polyhedra and its mirror image, joined at the
origin. One of the pointgroup symmetry operations should move the
inverted polyhedron to have a face in common with the first polyhedron,
which when combined correctly constructs a convex polyhedron.

The Polyhedron class in brille does not handle the union of two
polyhedra correctly where they overlap, share one edge, or share
opposing facets; but it does keep only unique vertices.
As a workaround the introduced algorithm looks for the combination of
the original polyhedron and its pointgroup-operated inversion which has
the smallest number of vertices. A convex hull of the points of these
unique vertices is constructed and, if it has the correct volume, it
must be a convex irreducible polyhedron and is stored as *the*
irreducible Brillouin zone.