[Update] repository location references and tutorial
References to the repository location have been updated to point to
https://github.com/brille/brille in line with its move to the new
organizational account.
The brilleu tutorial highlighting some of the utility of brille has been
modified to make use of a NaCl castep binary file stored, now, in the
brilleu repository. The automatic-fetching routines have been moved to
brilleu's utilities submodule as well.

The 'validation' script has also been moved to the brilleu repository
and is now configured to run automatically as an integration test.