[Fix] Polyhedron bisection bug
Bisecting a Polyhedron by a set of planes could (somehow) create
leftover vertices, normals, and points of a 'polyhedron' which only
has one plane.
A check has been added to ensure that there are at least 4 vertices,
4 normals, and 4 on-plane points otherwise the 'polyhedron' is
invalid and an empty Polyhedron() is returned.

Also simplified is Polyhedron::intersects by removing the recentring
of the reference Polyhedron and other Polyhedron planes in order to
avoid the Polyhedron copy constructor.