Improve PolyhedronTrellis creation time (#35)
* [Fix] Improvements for PolyhedronTrellis creation

Reimplement checks to decide whether the intersection between a node
and the input Polyhedron must be determined. Earlier versions of these
checks were not robust and eliminated nodes which intersected with the
Polyhedron. The new checks are tested against a case where the old
checks failed.

* [Fix] Polyhedron bisection bug

Bisecting a Polyhedron by a set of planes could (somehow) create leftover 
vertices, normals, and points of a 'polyhedron' which only has one plane.
A check has been added to ensure that there are at least 4 vertices, 4 normals,
and 4 on-plane points otherwise the 'polyhedron' is invalid and an empty 
Polyhedron() is returned.

Also simplified is Polyhedron::intersects by removing the recentring of the
reference Polyhedron and other Polyhedron planes in order to avoid the
Polyhedron copy constructor.