[Fix] GitHub Action on macOS host CMake install error
Letting pip install CMake causes an error due to the absence of
'skbuild' which appears to indicate that a wheel does not exist for
macOS on PyPI. Adding 'scikit-build' to the dependencies would fix the
issue but then CMake would be built from scratch.
Using brew to perform the install at least removes the dependency on the
scikit-build python package.