[Bug Fixes | Documentation Improvements]
Minor bug in parsing CIF xyz format strings fixed:
If the string did not end in a ';' or ',' character the parser
failed previously. Now it ensure that the last-read character is
valid before attempting to continue parsing.

Improvements to documentation have been made.
Added sections for `Grids` and `Symmetry` for the `_brille` module.
Building locally using the same Docker container as on Github Actions
is now supported -- the action now has an extra flag to control automatic
updating to the GH-pages repository/repositories.