[Fix] Undefined ssize_t
The type ssize_t is used by pybind11 to indicate a signed integer
similar to size_t, which is usually a 64-bit unsigned integer. The
pybind11 specific integer type exists in some libraries (gcc, for one)
independent of pybind11. For other libraries (Visual Studio) the type is
not defined and an error is thrown when it is encountered. Strangely VS
somehow found its definition previously but does not any longer.

To fix the missing type definition, all uses of ssize_t are modified to
either pybind11::ssize_t, to a templateded generic integer, or prefaced
by a `using pybind11::ssize_t;` statement.
This *should* fix the failed-build problem on the Windows GitHub Actions
worker, but is not yet tested fully.