Documentation (#45)
* Progress on documenting more of _brille

* _brille.BrillouinZone documentation updated

* [Bug Fixes | Documentation Improvements]

Minor bug in parsing CIF xyz format strings fixed:
If the string did not end in a ';' or ',' character the parser
failed previously. Now it ensure that the last-read character is
valid before attempting to continue parsing.

Improvements to documentation have been made.
Added sections for `Grids` and `Symmetry` for the `_brille` module.
Building locally using the same Docker container as on Github Actions
is now supported -- the action now has an extra flag to control automatic
updating to the GH-pages repository/repositories.

* Updates to documentation

- Add an offline builder using the same workflow as on Github Actions
- Add a second tutorial
- Update documentation throughout

* [Fix] Undefined ssize_t

The type ssize_t is used by pybind11 to indicate a signed integer
similar to size_t, which is usually a 64-bit unsigned integer. The
pybind11 specific integer type exists in some libraries (gcc, for one)
independent of pybind11. For other libraries (Visual Studio) the type is
not defined and an error is thrown when it is encountered. Strangely VS
somehow found its definition previously but does not any longer.

To fix the missing type definition, all uses of ssize_t are modified to
either pybind11::ssize_t, to a templateded generic integer, or prefaced
by a `using pybind11::ssize_t;` statement.
This *should* fix the failed-build problem on the Windows GitHub Actions
worker, but is not yet tested fully.