[Add] preprocessor macro to control HDF use
Getting HDF up and running in the CI system is not/will not be easy.
To avoid problems with this, all HDF interfaces are now controlled by a
single preprocessor macro 'USE_HIGHFIVE' with can be set by CMake if
'BRILLE_HDF' is 'ON'.
Otherwise, all HDF IO seems to be in place and working, with
BrillouinZone, BZTrellisQ* and BZNestQ* HDF IO exposed through Python.