[Fix] Error raised by test under MSVC
- One C++ test, PolyhedronTrellis construction from 'P1' hexagonal
system must contain Gamma, caused a segmentation fault when compiled
with the Microsoft Visual Studio compiler. The fault was raised when
attempting to de-reference the null pointer in a routine to identify
which tetrahedron contains the Gamma point. How the tetgen structure
ended up with a pointer (to a pointer ...) to 0x0 is not clear, but
likely is due to an uninitialized variable.
- The exact origin of the segmentation fault remains a mystery, but
changing from TetGen v1.5 to v1.6 has corrected the problem under MSVC
and Apple Clang compliers, at least. Along with this change it was
necessary to update the brille-TetGen interface where it appears in the
`triangulation*` files to explicitly set each polygon to have *no* holes
with the `holelist` pointer set to the `nullptr`, previously this was
implicitly set and the holelist pointer was uninitialized under v1.6.
- The release of TetGen v1.6.0 came in August 2020 without a changelog.
I have thus-far avoided switching from v1.5 due to this, and small
changes I've made to allow, e.g., geometry predicate computations with
constant point-list pointers. At this point there is still no reference
material for v1.6, and the only announcement about its release at
https://www.eurekalert.org/news-releases/755729
is not very informative. Hopefully the boasted enhancements are real and
worthwhile for users of brille.