[Fix] basis vector snap_to_symmetry
Two basis vectors related by a symmetry operation *must* have the same
length. This is true for real and reciprocal lattice basis vectors,
which are themselves related by a* = (2pi/V) b x c, and the two cyclic
permutations. So if we know that some subset of (a,b,c) and (a*,b*,c*)
must have the same lengths we can restrict both the lengths *and the
angles* of the real lattice basis vectors!