[Refactor] vertex map knitting to use maps
Running the interpolation volume polyhedron intersection test with the
node-cubes of a trellis in parallel is much faster than not, but sharing
a single vertex-mapping structure with concurrency locks on the vertices
seems to be a bottleneck still.
By giving each thread its own vertex-mapping structure the threads can
work independently through the list of all nodes, which moves the
challenge to re-combining all kept and appended vertices until after the
parallel region.
That recombination is complex and should be possible to do pairwise and
out of order, which would then work in parallel.

This commit makes improvements in the re-combining process that are
aimed at improving its speed and reliability. Instead of holding
N-length vectors for the N vertices of the input 'pristine' trellis
vertices, it now uses (trellis-index -> output-index) standard maps
which reduce the number of elements in each container *and* simplify
much of the logic.

This new approach works! But, strangely, only for the first
re-combination of mapping data. (That is, if there are two threads in
the parallel region it succeeds.) A first attempt to have more
re-combinations (from 12 -> 6 -> 3 -> 2 -> 1) failed for a thus unknown
reason.
More investigation is required, but this progress must be secured.