[Fix] working directory for setuptools_scm version information
All git and setuptools_scm calls should be performed *from* the root of
the project, which is `${CMAKE_SOURCE_DIR}` only
**durring the configuration**!
When the custom `AlwaysCheckGit` target is later called a compilation
the same environment variable points to the configured build directory,
which does not contain sufficient information for setuptools_scm and may
not provide any git information for, e.g., out-of-source builds.

The `AlwaysCheckGit` target *already* stores the first encountered value
of `${CMAKE_SOURCE_DIR}` as an envrionment variable,
`${pre_configure_dir}`, which is now used as the working directory for
the git and setuptools_scm process calls.
Other process calls no longer explicity set a working directory as they
are independent of where they are run.