[Update] setuptools config
Building on a 'win-amd64' host somehow produced 32-bit wheels via
	python -m pip wheel .
possibly due to use of an old-style build configuration in setup.py

The old configuration is updated with the latest pybind11 CMake example.
This includes checks for the running Windows platform, etc., so
hopefully the problem will also be fixed.

Along with this change, static setup information is moved into
pyproject.toml to minimize the project specific static data in setup.py